+++ srcs/ft_atoi.c
@@ -6,16 +6,16 @@
 /*   By: craffate <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/04 11:06:22 by craffate          #+#    #+#             */
-/*   Updated: 2016/11/24 22:09:11 by craffate         ###   ########.fr       */
+/*   Updated: 2017/02/24 10:07:07 by craffate         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_atoi(const char *str)
+long long int	ft_atoi(const char *str)
 {
 	unsigned int	i;
-	unsigned int	j;
+	long long int	j;
 	int				k;
 
 	i = 0;
